Create secret.xml template, replacing the file.
Simply listing 'secret.xml' in .gitignore does not work because the file
is still tracked by git. Only files that are untracked are covered by .gitignore.
So if 'secret.xml' was changed, it will still be picked up by git. If we
told git to untrack the file in order for .gitignore to work, then the file will be removed in the repo
entirely, which is not what we want either.

Instead, we rename 'secret.xml' to 'secret_template.xml' and use it as a template.
Developers then make a copy of that file and insert the API key there. The original template file remains
unchanged and still tracked by git, but the new file is what's actually
being used by the app. The new file will be listed under .gitignore and
we ask developers to not add the new file to git so git doesn't
track it.